refactor:[R6-trivial] Rename presentation layer to presenter
