+++ lib/presenter/bloc/movies/movie_state.dart
