+++ lib/presenter/bloc/series/series_bloc.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:r6_moovie_app/presentation/bloc/series/series_event.dart';
-import 'package:r6_moovie_app/presentation/bloc/series/series_state.dart';
+import 'package:r6_moovie_app/presenter/bloc/series/series_event.dart';
+import 'package:r6_moovie_app/presenter/bloc/series/series_state.dart';
 import 'package:r6_moovie_app/resources/app_strings.dart';
 import '../../../domain/entities/series.dart';
 import '../../../domain/usecase/get_popular_series_use_case.dart';
