+++ lib/presenter/bloc/series/series_event.dart
