+++ lib/presenter/bloc/series/series_state.dart
