+++ lib/presenter/widgets/details/arch_banner_image.dart
