+++ lib/presenter/widgets/details/info_row.dart
