+++ lib/presenter/widgets/details/media_detail_header.dart
