+++ lib/presenter/widgets/details/text_list.dart
