+++ lib/presenter/widgets/home/media_list.dart
@@ -1,9 +1,9 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:r6_moovie_app/presentation/pages/movies_details_screen.dart';
 import 'package:r6_moovie_app/resources/app_values.dart';
 import '../../../domain/entities/movie.dart';
 import '../../../domain/entities/series.dart';
+import '../../pages/movies_details_screen.dart';
 import '../../pages/series_details_screen.dart';
 
 class MediaList extends StatelessWidget {
